backend/requirements.txt
#LangChain Core
langchain 
langchain-core 

#OpenAI Integration
langchain-openai 
openai 

#Anthropic Integration
langchain-anthropic 

#Google Gemini Integration
langchain-google-genai 
google-generativeai 

#Hugging Face Integration
langchain-huggingface 
transformers 
huggingface-hub 
langchain-community

#Environment Variable Management
python-dotenv 

#Machine Learning Integration
numpy
scikit-learn
torch